refactor(dh): move em dash pipe to shared domain (#1269)
* refactor(dh): move em dash pipe to global shared domain

* refactor(dh): adjust import path to em dash

* refactor(dh): move em dash pipe to global shared domain

* chore: add license

* fix(dh): fix imports
