+++ libs/dh/market-participant/feature-gridarea/src/lib/overview/dh-market-participant-gridarea-overview.component.ts
@@ -36,7 +36,7 @@ import { WattEmptyStateModule } from '@energinet-datahub/watt/empty-state';
 import { WattButtonModule } from '@energinet-datahub/watt/button';
 import { WattBadgeComponent } from '@energinet-datahub/watt/badge';
 import { MatMenuModule } from '@angular/material/menu';
-import { DhEmDashFallbackPipeScam } from '@energinet-datahub/dh/metering-point/shared/ui-util';
+import { DhEmDashFallbackPipeScam } from '@energinet-datahub/dh/shared/ui-util';
 import {
   GridAreaChanges,
   GridAreaOverviewRow,
