+++ libs/dh/message-archive/feature-log-search/src/lib/drawer/dh-message-archive-drawer.component.ts
@@ -33,7 +33,7 @@ import { DhMessageArchiveDataAccessBlobApiStore } from '@energinet-datahub/dh/me
 import { ActorNamePipe } from '../shared/dh-message-archive-actor.pipe';
 import { DocumentTypeNamePipe } from '../shared/dh-message-archive-documentTypeName.pipe';
 import { PushModule } from '@rx-angular/template';
-import { DhEmDashFallbackPipeScam } from '@energinet-datahub/dh/metering-point/shared/ui-util';
+import { DhEmDashFallbackPipeScam } from '@energinet-datahub/dh/shared/ui-util';
 
 import { findLogName } from '../shared/findLogname';
