+++ libs/dh/metering-point/feature-identity-and-master-data/src/lib/identity/dh-metering-point-identity.component.ts
@@ -26,8 +26,8 @@ import { TranslocoModule } from '@ngneat/transloco';
 import { MeteringPointCimDto } from '@energinet-datahub/dh/shared/domain';
 import { DhMeteringPointIdentityTextFieldWithIconComponent } from './identity-text-field/dh-metering-point-identity-text-field-with-icon.component';
 import { DhMeteringPointStatusBadgeScam } from '@energinet-datahub/dh/metering-point/ui-status-badge';
-import { DhEmDashFallbackPipeScam } from '@energinet-datahub/dh/metering-point/shared/ui-util';
 import { DhSharedUiDateTimeModule } from '@energinet-datahub/dh/shared/ui-date-time';
+import { DhEmDashFallbackPipeScam } from '@energinet-datahub/dh/shared/ui-util';
 
 export interface MeteringPointIdentityTranslationKeys {
   meteringMethod: string;
