+++ libs/dh/metering-point/feature-identity-and-master-data/src/lib/primary-master-data/dh-metering-point-primary-master-data.component.ts
@@ -24,7 +24,7 @@ import { CommonModule } from '@angular/common';
 import { TranslocoModule } from '@ngneat/transloco';
 
 import { MeteringPointCimDto } from '@energinet-datahub/dh/shared/domain';
-import { DhEmDashFallbackPipeScam } from '@energinet-datahub/dh/metering-point/shared/ui-util';
+import { DhEmDashFallbackPipeScam } from '@energinet-datahub/dh/shared/ui-util';
 
 export type PrimaryMasterData = Pick<
   MeteringPointCimDto,
