+++ libs/dh/metering-point/feature-identity-and-master-data/src/lib/secondary-master-data/dh-secondary-master-data.component.spec.ts
@@ -29,7 +29,7 @@ import {
 import { getTranslocoTestingModule } from '@energinet-datahub/dh/shared/test-util-i18n';
 import { en as enTranslations } from '@energinet-datahub/dh/globalization/assets-localization';
 import { runOnPushChangeDetection } from '@energinet-datahub/dh/shared/test-util-metering-point';
-import { emDash } from '@energinet-datahub/dh/metering-point/shared/ui-util';
+import { emDash } from '@energinet-datahub/dh/shared/ui-util';
 
 import {
   DhSecondaryMasterDataComponent,
