+++ libs/dh/shared/ui-util/src/test-setup.ts
@@ -14,6 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { DhYesNoPipeScam } from './lib/dh-yes-no.pipe';
-export { DhIsParentPipeScam } from './lib/dh-is-parent.pipe';
-export { DhShowForMeteringPointTypeDirectiveScam } from './lib/dh-show-for-metering-point-type.directive';
+import 'jest-preset-angular/setup-jest';
+
+import { setUpTestbed } from '@energinet-datahub/gf/test-util-staging';
+
+setUpTestbed();
