+++ workspace.json
@@ -58,6 +58,7 @@
     "dh-shared-test-util-msw": "libs/dh/shared/test-util-msw",
     "dh-shared-ui-date-time": "libs/dh/shared/ui-date-time",
     "dh-shared-ui-paginator": "libs/dh/shared/ui-paginator",
+    "dh-shared-ui-util": "libs/dh/shared/ui-util",
     "dh-shared-util-application-insights": "libs/dh/shared/util-application-insights",
     "dh-shared-util-table": "libs/dh/shared/util-table",
     "dh-wholesale-data-access-api": "libs/dh/wholesale/data-access-api",
